Reorganize tests
1. Change the test pattern to be `test_*.rb` rather than `*_test.rb` to
   match nokogiri tests.
2. Split the encoding tests into their own file.
